requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
instructor>=1.0.0
langfuse>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pycaret>=3.3.0
boto3>=1.34.0
matplotlib>=3.8.0
openai>=1.30.0